How to Save Money and Still Have Fun as a Student
Over the past couple of decades, the mindset of the average citizen of the world has greatly changed. Our focus on things that make us happy has shifted towards consumerism. People around the world are always thinking about ways to earn a lot of money and become rich. Well, it all sounds nice; however, having a lot of money does not mean that you are going to be happy. Therefore, it is completely possible to cut down your expenses and still have fun with those surrounding you.

So, if you are a student who is not in a great financial situation, there are numerous ways in which you can save money and still not miss out on all the great things in this period of life.

##Attend Free Events
People believe that there are no free events around town or their campus, but think again. You can do some research and dig out a variety of fun places where people organize events that are completely free. This is a great activity for you and your friends – organize yourselves and rush to one of these events; you are going to be amazed at how much fun you are going to have. 
A little bonus lies in the fact that you never know who you might meet; maybe a future partner or a part-time owner who is going to offer you a job? Consider this activity as networking while you are having fun with your friends. You are going to love it! Plus, you get all this for free.
Organize time with friends differently
If you want to avoid the usual expenses, you should start avoiding expensive nights out. It might be fun at the moment, but your wallet is going to get thinner in no time, especially if you practice this kind of activity two or three times per week.
When you want to save some money, but not sacrifice your social life, you need to be creative and organize fun gatherings with your friends. Organize occasional walks downtown or hiking trips. A picnic is a great pass time activity that all of you can enjoy. Use it to get to know each other better. If your friends are into sports, you can organize dozens of fun physical activities, which are beneficial for both your body and your mind.
Make awesome meals
 Okay, this is the part where you can save a lot of money. If you are one of the people who already have their meals sorted out through a University organization, then you need to simply stop throwing away your money on snacks and useless sodas or other types of drinks. You will see how much money you can save by simply taking this step. However, if you are completely on your own, you should dedicate a little bit of time and learn how to prepare several meals.
With a variety of online recipe websites, this should be a fairly easy task – simply start researching and stop being lazy. Some people do not realize how much money they can save by avoiding fast food. Not only is it tremendously bad for your health, but it is also going to make you burn through your budget fast. This is why it is very important to organize yourself and dedicate your time to cooking. Within only a week, you are going to adjust to the new nutrition and your body is also going to love it.
Make a list when you go shopping
This is a mistake that almost everyone makes. You see that you need a couple of things in your home, you decide to go shopping and you return with a bunch of stuff that you actually didn’t need at all. Every time before you go to a store, make a list and stick to it. There will always be numerous products that will make you want to buy them. By making a list, you will know where every single cent went.
These are only some of the tips that you could use to save money as a student. There are dozens of other ways that you can use to save money while still having fun as a student. It might be hard to get used to it, but it will become a great habit that is going to turn you into a much more responsible person.